requirements.txt
absl-py==0.9.0
altgraph==0.17
astunparse==1.6.3
async-generator==1.10
backcall==0.2.0
cachetools==3.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dm-tree==0.1.5
entrypoints==0.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.10
ipython-genutils==0.2.0
Jinja2==2.11.2
jupyter==1.0.0
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
mpmath==1.1.0
nest-asyncio==1.4.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
parso==0.7.0
Pillow==8.0.1
portpicker==1.3.1
prometheus-client==0.8.0
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller-hooks-contrib==2020.10
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==19.0.2
QtPy==1.9.0
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scipy==1.5.4
semantic-version==2.8.5
Send2Trash==1.5.0
six==1.15.0
tb-nightly==2.4.0a20201112
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
tensorflow-federated-nightly==0.17.0.dev20201112
tensorflow-hub==0.10.0
tensorflow-model-optimization==0.4.1
tensorflow-privacy==0.5.1
tensorflowjs==2.7.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf-estimator-nightly==2.4.0.dev2020102301
tf-nightly==2.5.0.dev20201112
tfa-nightly==0.12.0.dev20201110202233
tornado==6.0.4
typed-ast==1.4.1
typeguard==2.10.0
typing-extensions==3.7.4.3
urllib3==1.26.1
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1